scathat-data-base/requirements.txt
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pinecone-client==3.2.2
numpy==1.24.3